int type
git-svn-id: svn+ssh://astrometry.net/svn/trunk/src/astrometry@18236 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ util/ioutils.c
@@ -481,7 +481,7 @@ static int readfd(int fd, char* buf, int NB, char** pcursor,
 	if (nr == 0) {
 		if (cursor != buf) {
 			//printf("flushing the last line\n");
-			sl_appendf(lines, "%.*s", (cursor - buf), buf);
+			sl_appendf(lines, "%.*s", (int)(cursor - buf), buf);
 		}
 		*pdone = TRUE;
 		return 0;
